pyproject.toml
[project]
name = "json-rpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.5",
    "pandas>=3.0.1",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.8",
    "jsonrpcserver>=5.0.9",
    "uvicorn>=0.34.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langchain-mcp-adapters>=0.2.1",
    "streamlit>=1.19.0",
    "logging>=0.4.9.6",
    "langchain-ollama>=1.0.1",
]
